+++ app/models/recurring_task.rb
@@ -19,8 +19,8 @@ class RecurringTask < ActiveRecord::Base
   validates_presence_of :interval_localized_name
   validates_presence_of :interval_number
   
-  validates :interval_localized_name, inclusion: { in: RecurringTask::INTERVAL_UNITS_LOCALIZED, message: "#{l(:error_invalid_interval)} '%{value}' (Validation)" }
-  validates :interval_number, numericality: {only_integer: true, greater_than: 0}
+  validates_inclusion_of :interval_localized_name, :in => RecurringTask::INTERVAL_UNITS_LOCALIZED, :message => "#{l(:error_invalid_interval)} '%{value}' (Validation)"
+  validates_numericality_of :interval_number, :only_integer => true, :greater_than => 0
   # cannot validate presence of issue if want to use other features; requiring presence of fixed_schedule requires it to be true
 
   validates_associated :issue # just in case we build in functionality to add an issue at the same time, verify the issue is ok