+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 
 * Using validates_presence_of instead of validates :x, presence: true for backward Rails compatibility (#20)
 * Using validates_inclusion_of and validates_numericality_of instead of stand-alone validates for backward Rails compatibility (#20)
+* Creating a new recurrence sets interval_unit properly from localized name
 
 ## Version 1.2.7 (04 Jan 2014)